decision/requirements.txt
# Decision Service Dependencies
pyyaml>=6.0.1
python-dotenv>=1.0.0
prometheus-client>=0.19.0
structlog>=24.1.0
orjson>=3.9.0
aiofiles>=23.2.1
